+++ src/puppetter.ts
@@ -9,10 +9,14 @@ export async function login(page, form) {
   }
 }
 
-export async function startWithUrl(puppeteer, url: string) {
+export async function startWithUrl(
+  puppeteer,
+  url: string,
+  options: { debug: boolean }
+) {
   try {
     const browser = await puppeteer.launch({
-      headless: process.env.NODE_ENV === 'development' ? false : true
+      headless: !options.debug
     });
     const page = await browser.newPage();
     const override = Object.assign(page.viewport(), { width: 1366 });